renaming + fix compile issue
