+++ contracts/test/Lib.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: BlueOak-1.0.0
 pragma solidity 0.8.9;
 
-import "contracts/generic/ProtoState.sol";
+import "contracts/test/ProtoState.sol";
 
 import "hardhat/console.sol";
