+++ tasks/p0/deploy_all.ts
@@ -1,39 +1,39 @@
-import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
-import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { bn, fp } from '../../common/numbers';
-import { RSRAssetP0 } from '../../typechain/RSRAssetP0';
-import { ERC20Mock } from '../../typechain/ERC20Mock';
-import { BigNumber, ContractFactory } from 'ethers';
+import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
+import { HardhatRuntimeEnvironment } from 'hardhat/types'
+import { bn, fp } from '../../common/numbers'
+import { RSRAssetP0 } from '../../typechain/RSRAssetP0'
+import { ERC20Mock } from '../../typechain/ERC20Mock'
+import { BigNumber, ContractFactory } from 'ethers'
 import { task } from 'hardhat/config'
-import { COMPAssetP0 } from '../../typechain/COMPAssetP0';
-import { AAVEAssetP0 } from '../../typechain/AAVEAssetP0';
-import { CompoundOracleMockP0 } from '../../typechain/CompoundOracleMockP0';
-import { ComptrollerMockP0 } from '../../typechain/ComptrollerMockP0';
-import { AaveOracleMockP0 } from '../../typechain/AaveOracleMockP0';
-import { AaveLendingAddrProviderMockP0 } from '../../typechain/AaveLendingAddrProviderMockP0';
-import { AaveLendingPoolMockP0 } from '../../typechain/AaveLendingPoolMockP0';
-import { VaultP0 } from '../../typechain/VaultP0';
-import { IManagerConfig, IParamsAssets } from '../../test/proto0/utils/fixtures';
-import { DeployerP0 } from '../../typechain/DeployerP0';
-import { expectInReceipt } from '../../common/events';
-import { MainP0 } from '../../typechain/MainP0';
-import { RTokenP0 } from '../../typechain/RTokenP0';
-import { FurnaceP0 } from '../../typechain/FurnaceP0';
-import { StRSRP0 } from '../../typechain/StRSRP0';
-import { DefaultMonitorP0 } from '../../typechain/DefaultMonitorP0';
-import { AssetManagerP0 } from '../../typechain/AssetManagerP0';
+import { COMPAssetP0 } from '../../typechain/COMPAssetP0'
+import { AAVEAssetP0 } from '../../typechain/AAVEAssetP0'
+import { CompoundOracleMockP0 } from '../../typechain/CompoundOracleMockP0'
+import { ComptrollerMockP0 } from '../../typechain/ComptrollerMockP0'
+import { AaveOracleMockP0 } from '../../typechain/AaveOracleMockP0'
+import { AaveLendingAddrProviderMockP0 } from '../../typechain/AaveLendingAddrProviderMockP0'
+import { AaveLendingPoolMockP0 } from '../../typechain/AaveLendingPoolMockP0'
+import { VaultP0 } from '../../typechain/VaultP0'
+import { IManagerConfig, IParamsAssets } from '../../test/p0/utils/fixtures'
+import { DeployerP0 } from '../../typechain/DeployerP0'
+import { expectInReceipt } from '../../common/events'
+import { MainP0 } from '../../typechain/MainP0'
+import { RTokenP0 } from '../../typechain/RTokenP0'
+import { FurnaceP0 } from '../../typechain/FurnaceP0'
+import { StRSRP0 } from '../../typechain/StRSRP0'
+import { DefaultMonitorP0 } from '../../typechain/DefaultMonitorP0'
+import { AssetManagerP0 } from '../../typechain/AssetManagerP0'
 
 const qtyHalf: BigNumber = bn('1e18').div(2)
 const qtyThird: BigNumber = bn('1e18').div(3)
 const qtyDouble: BigNumber = bn('1e18').mul(2)
 
 function waitDeployment(contracts: any[]) {
-  return Promise.all(contracts.map(contract => contract.deployed()))
+  return Promise.all(contracts.map((contract) => contract.deployed()))
 }
 
 const deployRSR = async (hre: HardhatRuntimeEnvironment, deployer: SignerWithAddress) => {
   const ERC20: ContractFactory = await hre.ethers.getContractFactory('ERC20Mock')
-  const rsr: ERC20Mock = <ERC20Mock> await ERC20.connect(deployer).deploy('Reserve Rights', 'RSR')
+  const rsr: ERC20Mock = <ERC20Mock>await ERC20.connect(deployer).deploy('Reserve Rights', 'RSR')
   await rsr.deployed()
 
   const RSRAssetFactory: ContractFactory = await hre.ethers.getContractFactory('RSRAssetP0')
@@ -43,7 +43,12 @@ const deployRSR = async (hre: HardhatRuntimeEnvironment, deployer: SignerWithAdd
 }
 
 const deployVault = async (hre: HardhatRuntimeEnvironment, deployer: SignerWithAddress) => {
-  const TOKENS = [['Token 0', 'TKN0'], ['Token 1', 'TKN1'], ['Token 2', 'TKN2'], ['Token 3', 'TKN3']]
+  const TOKENS = [
+    ['Token 0', 'TKN0'],
+    ['Token 1', 'TKN1'],
+    ['Token 2', 'TKN2'],
+    ['Token 3', 'TKN3'],
+  ]
   const ERC20: ContractFactory = await hre.ethers.getContractFactory('ERC20Mock')
 
   const vaultTokens = await Promise.all(TOKENS.map((tokenParams) => ERC20.connect(deployer).deploy(...tokenParams)))
@@ -52,10 +57,12 @@ const deployVault = async (hre: HardhatRuntimeEnvironment, deployer: SignerWithA
   // Set Collateral Assets and Quantities
   const AssetFactory: ContractFactory = await hre.ethers.getContractFactory('CollateralP0')
   // Deploy vault tokens as collaterals
-  const collaterals = await Promise.all(vaultTokens.map((token) => AssetFactory.connect(deployer).deploy(token.address, token.decimals())))
+  const collaterals = await Promise.all(
+    vaultTokens.map((token) => AssetFactory.connect(deployer).deploy(token.address, token.decimals()))
+  )
   await waitDeployment(collaterals)
 
-  const collateral: string[] = collaterals.map(c => c.address)
+  const collateral: string[] = collaterals.map((c) => c.address)
   const quantities: BigNumber[] = [qtyHalf, qtyHalf, qtyThird, qtyDouble]
 
   const VaultFactory: ContractFactory = await hre.ethers.getContractFactory('VaultP0')
@@ -82,17 +89,23 @@ const deployDependencies = async (hre: HardhatRuntimeEnvironment, deployer: Sign
 
   // Deploy Comp and Aave Oracle Mocks
   const CompoundOracleMockFactory: ContractFactory = await hre.ethers.getContractFactory('CompoundOracleMockP0')
-  const compoundOracle: CompoundOracleMockP0 = <CompoundOracleMockP0>await CompoundOracleMockFactory.connect(deployer).deploy()
+  const compoundOracle: CompoundOracleMockP0 = <CompoundOracleMockP0>(
+    await CompoundOracleMockFactory.connect(deployer).deploy()
+  )
   await compoundOracle.deployed()
 
   const ComptrollerMockFactory: ContractFactory = await hre.ethers.getContractFactory('ComptrollerMockP0')
-  const compoundMock: ComptrollerMockP0 = <ComptrollerMockP0>await ComptrollerMockFactory.connect(deployer).deploy(compoundOracle.address)
+  const compoundMock: ComptrollerMockP0 = <ComptrollerMockP0>(
+    await ComptrollerMockFactory.connect(deployer).deploy(compoundOracle.address)
+  )
   await compoundMock.deployed()
 
   const AaveOracleMockFactory: ContractFactory = await hre.ethers.getContractFactory('AaveOracleMockP0')
   const weth: ERC20Mock = <ERC20Mock>await ERC20.connect(deployer).deploy('Wrapped ETH', 'WETH')
   await weth.deployed()
-  const aaveOracle: AaveOracleMockP0 = <AaveOracleMockP0>await AaveOracleMockFactory.connect(deployer).deploy(weth.address)
+  const aaveOracle: AaveOracleMockP0 = <AaveOracleMockP0>(
+    await AaveOracleMockFactory.connect(deployer).deploy(weth.address)
+  )
   await aaveOracle.deployed()
 
   const AaveAddrProviderFactory: ContractFactory = await hre.ethers.getContractFactory('AaveLendingAddrProviderMockP0')
@@ -110,8 +123,7 @@ const deployDependencies = async (hre: HardhatRuntimeEnvironment, deployer: Sign
   return { weth, compToken, compAsset, compoundOracle, compoundMock, aaveToken, aaveAsset, aaveOracle, aaveMock }
 }
 
-task('Proto0-deployAll', 'Deploys all proto0 contracts and a mock RToken')
-.setAction(async (params, hre) => {
+task('Proto0-deployAll', 'Deploys all p0 contracts and a mock RToken').setAction(async (params, hre) => {
   const [deployer] = await hre.ethers.getSigners()
   // RSR
   console.log('Deploying RSR...')
@@ -122,7 +134,7 @@ task('Proto0-deployAll', 'Deploys all proto0 contracts and a mock RToken')
   // Dependencies
   console.log('Deploying Aave/Compound/Oracle dependencies')
   const { weth, compToken, compAsset, compoundOracle, compoundMock, aaveToken, aaveAsset, aaveOracle, aaveMock } =
-  await deployDependencies(hre, deployer)
+    await deployDependencies(hre, deployer)
 
   console.log('Setting prices...')
   // Set Default Oracle Prices
@@ -133,7 +145,7 @@ task('Proto0-deployAll', 'Deploys all proto0 contracts and a mock RToken')
   await compoundOracle.setPrice('ETH', bn('1e6'))
   await compoundOracle.setPrice('COMP', bn('1e6'))
 
-  Promise.all(vaultTokens.map(tkn => aaveOracle.setPrice(tkn.address, bn('1e18'))))
+  Promise.all(vaultTokens.map((tkn) => aaveOracle.setPrice(tkn.address, bn('1e18'))))
   await aaveOracle.setPrice(weth.address, bn('1e18'))
   await aaveOracle.setPrice(aaveToken.address, bn('1e18'))
   await aaveOracle.setPrice(compToken.address, bn('1e18'))
@@ -182,7 +194,7 @@ task('Proto0-deployAll', 'Deploys all proto0 contracts and a mock RToken')
       compoundMock.address,
       aaveMock.address,
       paramsAssets,
-      collaterals.map(c => c.address)
+      collaterals.map((c) => c.address)
     )
   ).wait()
