+++ test/p1/FurnaceP1.test.ts
